+++ kancolle/__init__.py
@@ -2,7 +2,6 @@ __author__ = 'me'
 
 from flask.ext.security import UserMixin, RoleMixin
 from db import db
-
 from kancolle.auth import *
-from kancolle.navalbase import *
-from kancolle.admiral import *
+from kancolle.classes import *
+from kancolle.api import *
